requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
environ==1.0
filelock==3.8.2
huggingface-hub==0.11.1
idna==3.4
kafka-python==2.0.2
numpy==1.24.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==22.0
psycopg2-binary==2.9.5
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
tokenizers==0.13.2
torch==1.13.1
tqdm==4.64.1
transformers==4.25.1
typing-extensions==4.4.0
urllib3==1.26.13
